+++ src/main/java/kitchenpos/domain/order/OrderLineItem.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import kitchenpos.common.BaseDate;
